// src/modules/marketplace/enums/order-status.enum.ts
// src/modules/marketplace/enums/order-status.enum.ts
export enum OrderStatus {
  BROUILLON = 'BROUILLON',
  ENVOYEE = 'ENVOYEE',
  ACCEPTEE = 'ACCEPTEE',
  REFUSEE = 'REFUSEE',
  PAYEE = 'PAYEE',
  EN_PREPARATION = 'EN_PREPARATION',
  EN_LIVRAISON = 'EN_LIVRAISON',
  LIVREE = 'LIVREE',
  ANNULEE = 'ANNULEE',
  TERMINEE = 'TERMINEE',
}